refactor: Fix login page to component
