+++ src/pages/Home.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState, useRef, useCallback, useMemo } from "react";
 import { IPost } from "../@types/posts";
 import { getPosts } from "../api/posts";
 import Postcard from "../components/Postcard";
-import Loading from "./Loading";
+import Loading from "../components/Loading";
 
 export default function Home() {
   const [posts, setPosts] = useState<Array<IPost>>([]);
